Initial working playback of one song

+++ app/src/main/java/com/example/homemusicplayer/data/apiResponse/mediaTypes/attributes/SongAttributes.kt
@@ -3,5 +3,5 @@ package com.example.homemusicplayer.data.apiResponse.mediaTypes.attributes
 class SongAttributes(
     val albumName: String,
     val artistName: String,
-    val durationInMillis: Int,
+    val durationInMillis: Long,
 ) : TypeAttributes()
